response: move CqlResponseKind to response module
It should not belong to errors module.
